fix: Cover rules including imported payees. (#70)
Implements the imported_payee type that was forgotten because it implements the exact same behaviour as string. There must be a reason why this was implemented, but as of now it implements exacly what STRING implements.

Also trimmed the imported_payee as the original API does, and set it to the payee if it's missing.