fix: Floating-Point Representation and Precision Error in transaction amount calculation (#77)
The issue is related to floating-point arithmetic in Python, which affects the accuracy of transaction amount calculations. Specifically, when performing arithmetic operations with floating-point numbers, the results can sometimes be slightly off due to the way floating-point numbers are represented in binary format and when these transactions accumulate, the final account budget result is different from the reality(sometimes by a significant amount). This can lead to incorrect financial tracking and budgeting

Example:

For instance, the calculation -9.95 * 100 results in -994.9999999999999 instead of the expected -995 and results in amount of -9.94 added to ActualBudget instead of -9.95.