perf: Improve performance of API queries by using the same session for all (#84)
Due to Actual rate limiting, the requests are throttled if coming from "different clients". This means that using the same session for all the requests make them faster.

Here is a benchmark before and after changes, for exporting a bit budget with many changes:

- Before changes: 35.13s
- After changes: 10.90s

This should only affect larger budgets that are remote.