fix: Allow creating transactions without payee (#89)
actualpy fails to create a transaction without a payee.

When a default (payee='') payee value is used, actualpy searches for a payee with an empty name and expects to find only one such payee. Actual Budget uses special payees to represent transfers between accounts. Such payees always have an empty name. If a budget has more than one transfer payee, create_transaction will fail because get_or_create_payee will find multiple payees with empty names and fail validation.

When a None payee value is passed, create_transaction succeeds, but then Actual errors because of a invalid payee. This change makes sure that the transaction is created with None payee_id and set it as the default, preventing the creation of an empty string payee and using the same default behaviour as Actual frontend.

---------

Co-authored-by: Brunno Vanelli <brunnovanelli@gmail.com>