fix: Adjust model for Balance.reference_date to be None if not passed (field is optional) (#100)
Fix model for Balance.reference_date to be None if not passed (optional). The  from gocardless seems to be missing the "referenceDate" in the two elements in "balances", as this item seems to be optional for some transactions.

This can be confirmed on their official API: https://developer.gocardless.com/bank-account-data/balance

---------

Co-authored-by: Brunno Vanelli <bvanelli@users.noreply.github.com>