feat: Handle balances when using Bank Sync (#108)
Allows the accounts to be reconciled when doing the first bank sync. This happens when the expected balance of the account does not match the final balance of the imported transactions, because old transactions are not imported.

This should only affect the first import using the bank sync, or if all transactions are deleted and a new bank sync is done.

Either way, it manages to reproduce 1:1 what we find on the simpleFin demo dataset, when importing from Actual or actualpy.

Closes #29