fix: Fix Bank sync sorting and initial value (#118)
Fixes the incorrect sorting of the Actual transactions since the resolution of actualbudget/actual-server#556

Since the server and frontend code have been moved together, all the bank sync validations are now being done on backend, to allow support for more integrations. This means actualpy has to handle less logic when evaluating the bank sync.

This PR solves a couple of issues:

- Correctly calculates the starting balance for goCardless integration
- Correctly sorts the transactions by starting with the "Starting Balance", then consuming all sync transactions in reverse
- Correctly set the sort_order field by multiplying it by 1000 to convert to milliseconds
- Added Accounts.last_sync and Transactions.raw_synced_data. Nothing has been implemented for them