feat: Make sure the tmp folder is consistent when launching the application multiple times. (#123)
When launching the application multiple times, the storage folder is always picked at random if not provided.

The PR makes sure that the uuid will be extracted from the metadata.json and then used to create a path that does not require the budget to be redowloaded every time, making the application faster when a storage folder is not provided but the script is called multiple times (i.e. development scenario).

Closes #122