feat: Implement changeset handler with docs. (#136)
The Changeset includes multiple column changes for an object modified on a remote server via frontend or API. The changeset contains the table that was modified, the unique id of the object as well as a dictionary of column changes. With this object, it is possible to retrieve the original value from the database for easier handling via the from_orm method.

It's important to note that the changeset does not contain the original values of the columns, as well as the information if the row is new or updated. It is possible, however, to retrieve this information by loading a local copy of the database.

The changeset is also only available from the moment the budget was initialized.

This PR contains one example on the examples folder on how to use the change handler.

Closes #135