fix: Only clear transaction category for transfers which are off-budget (#159)
When a rule marks a transaction as a transfer, the category is currently always set to None. This is correct for on-budget transfers, but off-budget transfers require a category. Examples include payments to loans or contributions to long-term savings accounts that are tracked as off-budget accounts in Actual.

In this case, the category for the on-budget transaction can be set via UI.

---------

Co-authored-by: Linus W. <github@mumindalen.net>
Co-authored-by: Brunno Vanelli <brunnovanelli@gmail.com>