tests: Use the pytest tmp_path fixture instead of tempfile for hosting temporary testing data (#166)
Currently the module tempfile is used for creating a tempfile for the SQLLite used for tests. This fails on windows due to file locking resulting in all test cases failing.

There is a built-in fixture in pytest which creates structured temporary directories and alse purges them automatically. This PR replaces the previous tempfile structure with the pytest fixture. See https://docs.pytest.org/en/stable/how-to/tmp_path.html for more information.

With this fix most of the test cases works as-is on Windows, I will go through the ones that still fail to see if they can be fixed as well.

Co-authored-by: Linus W. <github@mumindalen.net>